use run_path to retreive class name instead of ast
